Find by ip and by name
Try to find both CDP and LLDP by name and by ip.

If we find by name and find IP for it, prefer that. If we can't find by
name use peer address found by name.
These are different things as they are from separate TLV. Name is much
more likely to be useful, but name may not resolve to IP.